added dialogue interface
